Update create author create view fields.
The previous create author view fields were set to a vlaue of __all__, however this is not recommended due to potential security issues if more fields are added to the Author Model in future.

By explicitly declaring fields int he create Author view, we control which fields from the model are displayed.