Add language field to Book models
All books should have a language associated with them, thus the Book model requires a language field; this was previously missing.

In addition to associating each book record with a corresponding language, the language field will now be displayed on 'book detail' pages, and will be available in the 'create book' form.