Update DeleteBook view function and template.
The view function previously assigned the variable name 'e' to 'Exception' but 'e' was never used; this cuase Ruff to raise a problem. Thus 'e' was been removed.

The book_confirm_delete template was updated to show the book instance unique id next to each existing book instance. Given the titles are the same it's helpful to have a unique piece of information to distingush between copies.