Update author model
During MDN tutorial Part 10 (testing), a test was added to the unit tests for the author model. This test intterrogated the label for the date_of_death field which should have been added during creation of the Author model. I had missed this, and so needed to go back and add the field label, then make the necassary migrations.